Refactor Xliff Parsing for views
This commit adds the file XliffParser.php which is used to print
Xliff info to the screen. This affects the xview.php and xlifview.php
files. It takes the printing logic out of those two files and puts
it into a class for reuse. It also makes some changes to the logic.